Changed error handling to not use exceptions

Closes #12

See merge request cse1105/2022-2023/teams/oopp-team-37!29

+++ server/src/main/java/server/api/controllers/BoardController.java
@@ -18,8 +18,6 @@ package server.api.controllers;
 import commons.Board;
 import commons.CardList;
 import org.springframework.http.HttpStatus;
-import org.springframework.orm.jpa.JpaObjectRetrievalFailureException;
-import org.springframework.web.server.ResponseStatusException;
 import server.services.TextService;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
@@ -37,9 +35,10 @@ public class BoardController {
 
     /**
      * RestAPI Controller for the board route
-     * @param boardRepo repository for boards
+     *
+     * @param boardRepo          repository for boards
      * @param cardListRepository repository for cards
-     * @param textService service for generating random keys
+     * @param textService        service for generating random keys
      */
     public BoardController(final BoardRepository boardRepo,
                            final CardListRepository cardListRepository,
@@ -51,6 +50,7 @@ public class BoardController {
 
     /**
      * Handler for creating board
+     *
      * @param request the board payload
      * @return the created board
      */
@@ -63,63 +63,60 @@ public class BoardController {
 
     /**
      * Handler for getting the board
+     *
      * @param id the board id
      * @return the board
      */
     @GetMapping("/{id}")
     public ResponseEntity<Board> getBoard(@PathVariable final Integer id) {
-        try {
-            return new ResponseEntity<>(this.boardRepo.getById(id), new HttpHeaders(), 200);
-        } catch (final JpaObjectRetrievalFailureException e) {
-            throw new ResponseStatusException(
-                    HttpStatus.NOT_FOUND, "Board not found", e);
+        if (!this.boardRepo.existsById(id)) {
+            return ResponseEntity.notFound().build();
         }
+
+        return new ResponseEntity<>(this.boardRepo.getById(id), new HttpHeaders(), 200);
     }
 
     /**
      * Handler for creating the list in a board
+     *
      * @param id the board to create a list for
      * @return the board with its new list
      */
     @PostMapping("/{id}/list")
     public ResponseEntity<Board> createList(@PathVariable final Integer id) {
-        try {
-            final Board board = this.boardRepo.getById(id);
-
-            final CardList list = new CardList("New List");
-            this.cardListRepository.save(list);
-
-            board.addList(list);
-            this.boardRepo.save(board);
-            return new ResponseEntity<>(this.boardRepo.getById(id), new HttpHeaders(), 200);
-        } catch (final JpaObjectRetrievalFailureException e) {
-            throw new ResponseStatusException(
-                    HttpStatus.NOT_FOUND, "Board not found", e);
+        if (!this.boardRepo.existsById(id)) {
+            return ResponseEntity.notFound().build();
         }
+        final Board board = this.boardRepo.getById(id);
+
+        final CardList list = new CardList("New List");
+        this.cardListRepository.save(list);
+
+        board.addList(list);
+        return new ResponseEntity<>(this.boardRepo.save(board), new HttpHeaders(), 200);
     }
 
     /**
      * Handler for deleting the list of a board
-     * @param id the board id
+     *
+     * @param id     the board id
      * @param listId the list id to delete
      * @return the board without the list
      */
     @DeleteMapping("/{id}/list/{listId}")
     public ResponseEntity<Board> deleteList(@PathVariable final Integer id,
                                             @PathVariable final Integer listId) {
-        try {
-            final Board board = this.boardRepo.getById(id);
-            if (!board.removeListById(listId)) {
-                throw new ResponseStatusException(
-                        HttpStatus.NOT_FOUND, "List not found");
-            }
-
-            this.cardListRepository.deleteById(listId);
-            return new ResponseEntity<>(board, new HttpHeaders(), 200);
-        } catch (final JpaObjectRetrievalFailureException e) {
-            throw new ResponseStatusException(
-                    HttpStatus.NOT_FOUND, "Board not found", e);
+        if (!this.boardRepo.existsById(id)) {
+            return ResponseEntity.notFound().build();
         }
+
+        final Board board = this.boardRepo.getById(id);
+        if (!board.removeListById(listId)) {
+            return ResponseEntity.notFound().build();
+        }
+
+        this.cardListRepository.deleteById(listId);
+        return new ResponseEntity<>(this.boardRepo.save(board), new HttpHeaders(), HttpStatus.OK);
     }
 
 }
